-fixed some VS compiler warnings -added svn properties eol-style=native and mime-type=text/plain to OutboundTargetHandler
git-svn-id: https://svn.resiprocate.org/rep/resiprocate@7538 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ main/repro/monkeys/LocationServer.cxx
@@ -63,7 +63,7 @@ LocationServer::process(RequestContext& context)
                // does). Normally, this means we will ignore mReceivedFrom, but
                // this can be configured to still use mReceivedFrom.
                Target* target = new Target(contact);
-               target->mPriorityMetric=contact.mLastUpdated;
+               target->mPriorityMetric=(int)contact.mLastUpdated;
                outboundBatch[contact.mInstance].push_back(target);
             }
             else
